Cropped PDF before converting to an image.
I have utilized the pypdf library to crop the PDF from the full page to the floor plan. The PDF is cropped before being converted to an image so that only the relevant portion of the PDF is converted to an image. This allows for the image to be high quality and take less space.

i enabled use_pdftocairo when utilizing pdf2image's convert_from_path. This fixed the issue of losing walls when converting from PDF to image.